update laundry layout, average dryer and washer usage, fix laundry open status
