+++ PennMobile/src/main/res/values/dimens.xml
@@ -2,5 +2,6 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
-    <dimen name="laundry_machine_view">125dp</dimen>
+    <dimen name="laundry_machine_view">100dp</dimen>
+    <dimen name="laundry_font_small">14sp</dimen>
 </resources>
